+++ lib/helpers.js
@@ -25,3 +25,19 @@ export function callCmd(cmd, name, bot, msg, suffix) {
   console.log(`${chalk.blue('[' + moment().format('HH:mm:ss' + ']'))} ${chalk.bold.green(name)}: ${suffix}`);
   cmd(bot, msg, suffix);
 }
+
+export function getOrdinal(n) {
+  let s = ['th', 'st', 'nd', 'rd'];
+  let v = n % 100;
+  return n + (s[(v - 20) % 10] || s[v] || s[0]);
+}
+
+export function toTitleCase(str) {
+  return str.replace(/\w\S*/g, (txt) => {
+    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
+  });
+}
+
+export function numberWithCommas(x) {
+  return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+}
