+++ lib/leagueoflegends.js
@@ -4,6 +4,7 @@ import R from 'ramda';
 import _request from 'request';
 
 import sentry from './config/sentry';
+import { getOrdinal, toTitleCase } from './helpers';
 
 const request = Promise.promisify(_request);
 
@@ -50,18 +51,6 @@ function _makeRequest(options) {
     });
 }
 
-function _getOrdinal(n) {
-  let s = ['th', 'st', 'nd', 'rd'];
-  let v = n % 100;
-  return n + (s[(v - 20) % 10] || s[v] || s[0]);
-}
-
-function toTitleCase(str) {
-  return str.replace(/\w\S*/g, (txt) => {
-    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-  });
-}
-
 function counters(bot, msg, suffix) {
   let query = suffix.toLowerCase().replace('counters ', '').replace('counter ', '');
   let position = R.last(query.split(' '));
@@ -85,7 +74,7 @@ function counters(bot, msg, suffix) {
     .then(R.sortBy(R.prop('statScore')))
     .then(R.slice(0, 10))
     .then(data => {
-      return R.addIndex(R.map)((champ, idx) => `*${_getOrdinal(idx + 1)}*. **${champ.key}** - ${100 - champ.winRate}% win rate over ${champ.games} games.`, data);
+      return R.addIndex(R.map)((champ, idx) => `*${getOrdinal(idx + 1)}*. **${champ.key}** - ${100 - champ.winRate}% win rate over ${champ.games} games.`, data);
     })
     .then(R.prepend(`Sure! Here's the top 10 **statistical** counters for **${toTitleCase(champ)} ${gg_position}**:\n`))
     .then(R.join('\n'))
@@ -107,7 +96,7 @@ function _itemNames(item_data) {
     .then(R.path(['body', 'data']))
     .then(data => R.zipObj(R.keys(data), R.pluck('name')(R.values(data))))
     .then(items_list => {
-      item_data.item_names = R.addIndex(R.map)((item_id, idx) => `*${_getOrdinal(idx + 1)}*. ${items_list[item_id]}`, item_data.items);
+      item_data.item_names = R.addIndex(R.map)((item_id, idx) => `*${getOrdinal(idx + 1)}*. ${items_list[item_id]}`, item_data.items);
 
       return item_data;
     });
@@ -205,7 +194,7 @@ function bans(bot, msg) {
     .then(R.uniq)
     .then(R.slice(0, 10))
     .then(champs => {
-      return R.addIndex(R.map)((champ, idx) => `*${_getOrdinal(idx + 1)}*. **${toTitleCase(champ)}**`, champs);
+      return R.addIndex(R.map)((champ, idx) => `*${getOrdinal(idx + 1)}*. **${toTitleCase(champ)}**`, champs);
     })
     .then(R.prepend(`You got it! Here's the top 10 most common bans:\n`))
     .then(R.join('\n'))
@@ -234,7 +223,7 @@ function best(bot, msg, suffix) {
   return _makeRequest(options)
     .then(R.prop('data'))
     .then(data => {
-      return R.addIndex(R.map)((champ, idx) => `*${_getOrdinal(idx + 1)}*. **${champ.name}** with a ${champ.general.winPercent}% winrate.`, data);
+      return R.addIndex(R.map)((champ, idx) => `*${getOrdinal(idx + 1)}*. **${champ.name}** with a ${champ.general.winPercent}% winrate.`, data);
     })
     .then(R.prepend(`Sick! Here's the top 10 **statistically** best for **${gg_position}**:\n`))
     .then(R.join('\n'))
